pyproject.toml
[project]
name = "users-manager-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladyslav",email = "vlados.soc@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.1"
flask-swagger-ui = "^4.11.1"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.8.1"
email-validator = "^2.2.0"
flask-migrate = "^4.1.0"
pytest = "^8.3.5"
flask-testing = "^0.8.1"




[tool.poetry.group.dev.dependencies]
dependencies = "^7.7.1"
black = "^25.1.0"
ruff = "^0.11.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80

target-version = "py312"
select = [
    "B", "C", "E", "F", "W", "B9", "ANN", "Q0", "N8",
]
ignore = [
    "E203",
    "E266",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "ANN401",
    "N807",
    "N818",
    "F401",
    "B904"
]

exclude = [
    "migrations/*", "tests/*"
]